Moving widget folder/namespace inside viz folder/namespace
