+++ examples/_debug/dataview/category-aggregated-count-geojson.html
@@ -100,7 +100,7 @@
           mode: 'local'
         });
 
-        new carto.widget.Category('as-category-widget', dataview);
+        new carto.viz.widget.Category('as-category-widget', dataview);
       }
 
       initialize();
