+++ examples/dataview/category-data-view-all-modes.html
@@ -121,9 +121,9 @@
         { operation: 'sum', operationColumn: 'pop_est', mode: 'local' }
       );
 
-      new carto.widget.Category('#category-widget-remote', categoryRemote);
-      new carto.widget.Category('#category-widget-remote-filtered', categoryRemoteFiltered);
-      new carto.widget.Category('#category-widget-local', categoryLocal);
+      new carto.viz.widget.Category('#category-widget-remote', categoryRemote);
+      new carto.viz.widget.Category('#category-widget-remote-filtered', categoryRemoteFiltered);
+      new carto.viz.widget.Category('#category-widget-local', categoryLocal);
 
       // bind events
       categoryRemote.on('dataUpdate', async () => {
