+++ examples/widgets/category.html
@@ -90,7 +90,7 @@
           operation: 'count'
         });
 
-        new carto.widget.Category('as-category-widget', dataview);
+        new carto.viz.widget.Category('as-category-widget', dataview);
       }
 
       initialize();
