+++ examples/widgets/formula.html
@@ -88,7 +88,7 @@
           operation: 'max'
         });
 
-        new carto.widget.Formula('as-formula-widget', formulaDataview);
+        new carto.viz.widget.Formula('as-formula-widget', formulaDataview);
       }
 
       initialize();
