+++ src/lib/viz/widget/category/CategoryWidget.ts
